add Capture_Ether's Predict blockhash & forge fmt

+++ src/Capture_the_Ether/Lotteries/Guess_the_random_number/GuessTheRandomNumberChallenge.sol
@@ -1,4 +1,4 @@
-// SPDX-License-Identifier: GPL-3.0-or-later
+// SPDX-License-Identifier: MIT
 
 pragma solidity ^0.8.19;
 
@@ -7,7 +7,7 @@ contract GuessTheRandomNumberChallenge {
 
     constructor() payable {
         require(msg.value == 1 ether);
-        answer = uint8(uint(keccak256(abi.encodePacked(blockhash(block.number - 1), block.timestamp))));
+        answer = uint8(uint256(keccak256(abi.encodePacked(blockhash(block.number - 1), block.timestamp))));
     }
 
     function isComplete() public view returns (bool) {
